More tests and address comment
