+++ pkg/async/cloudevent/implementations/cloudevent_publisher.go
@@ -1,19 +1,17 @@
-package cloudevent
+package implementations
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"reflect"
 	"time"
+
+	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/admin"
 
 	"github.com/flyteorg/flyteadmin/pkg/async/notifications/implementations"
 
-	"github.com/Shopify/sarama"
-	"github.com/cloudevents/sdk-go/protocol/kafka_sarama/v2"
-
-	"github.com/NYTimes/gizmo/pubsub"
-	pbcloudevents "github.com/cloudevents/sdk-go/binding/format/protobuf/v2"
 	cloudevents "github.com/cloudevents/sdk-go/v2"
-	"github.com/google/uuid"
 
 	"k8s.io/apimachinery/pkg/util/sets"
 
@@ -25,63 +23,21 @@ import (
 )
 
 const (
-	cloudEventSource = "https://github.com/flyteorg/flyteadmin"
-	jsonSchemaURLKey = "jsonschemaurl"
-	jsonSchemaURL    = "https://github.com/flyteorg/flyteidl/blob/cloudevent2/jsonschema/workflow_execution.json"
+	cloudEventSource     = "https://github.com/flyteorg/flyteadmin"
+	cloudEventTypePrefix = "com.flyte.resource"
+	jsonSchemaURLKey     = "jsonschemaurl"
+	jsonSchemaURL        = "https://github.com/flyteorg/flyteidl/blob/cloudevent2/jsonschema/workflow_execution.json"
 )
-
-type Receiver = string
-
-const (
-	Kafka Receiver = "Kafka"
-)
-
-// Sender Defines the interface for sending cloudevents.
-type Sender interface {
-	// Send a cloud event to other services (pub/sub or Kafka)
-	Send(ctx context.Context, notificationType string, event cloudevents.Event) error
-}
-
-// PubSubSender Implementation of Sender
-type PubSubSender struct {
-	Pub pubsub.Publisher
-}
-
-func (s *PubSubSender) Send(ctx context.Context, notificationType string, event cloudevents.Event) error {
-	eventByte, err := pbcloudevents.Protobuf.Marshal(&event)
-	if err != nil {
-		logger.Errorf(ctx, "Failed to marshal cloudevent with error: %v", err)
-		return err
-	}
-	if err := s.Pub.PublishRaw(ctx, notificationType, eventByte); err != nil {
-		logger.Errorf(ctx, "Failed to publish a message with key [%s] and message [%s] and error: %v", notificationType, event.String(), err)
-		return err
-	}
-
-	return nil
-}
-
-// KafkaSender Implementation of Sender
-type KafkaSender struct {
-	Client cloudevents.Client
-}
-
-func (s *KafkaSender) Send(ctx context.Context, notificationType string, event cloudevents.Event) error {
-	if result := s.Client.Send(
-		// Set the producer message key
-		kafka_sarama.WithMessageKey(context.Background(), sarama.StringEncoder(event.ID())),
-		event,
-	); cloudevents.IsUndelivered(result) {
-		return fmt.Errorf("failed to send cloud event: %v", result)
-	}
-	return nil
-}
 
 // Publisher This event publisher acts to asynchronously publish workflow execution events.
 type Publisher struct {
 	sender        Sender
 	systemMetrics implementations.EventPublisherSystemMetrics
 	events        sets.String
+}
+
+func isInstanceOf(objectPtr, typePtr interface{}) bool {
+	return reflect.TypeOf(objectPtr) == reflect.TypeOf(typePtr)
 }
 
 func (p *Publisher) Publish(ctx context.Context, notificationType string, msg proto.Message) error {
@@ -91,22 +47,47 @@ func (p *Publisher) Publish(ctx context.Context, notificationType string, msg pr
 	p.systemMetrics.PublishTotal.Inc()
 	logger.Debugf(ctx, "Publishing the following message [%+v]", msg)
 
+	var executionId string
+	var phase string
+	var eventTime time.Time
+
+	switch reflect.TypeOf(msg) {
+	case reflect.TypeOf(admin.WorkflowExecutionEventRequest{}):
+		e := msg.(*admin.WorkflowExecutionEventRequest).Event
+		executionId = e.ExecutionId.String()
+		phase = e.Phase.String()
+		eventTime = e.OccurredAt.AsTime()
+	case reflect.TypeOf(admin.TaskExecutionEventRequest{}):
+		e := msg.(*admin.TaskExecutionEventRequest).Event
+		executionId = e.TaskId.String()
+		phase = e.Phase.String()
+		eventTime = e.OccurredAt.AsTime()
+	case reflect.TypeOf(admin.NodeExecutionEventRequest{}):
+		e := msg.(*admin.NodeExecutionEventRequest).Event
+		executionId = msg.(*admin.NodeExecutionEventRequest).Event.Id.String()
+		phase = e.Phase.String()
+		eventTime = e.OccurredAt.AsTime()
+	default:
+		return errors.New(fmt.Sprintf("Unsupported event types [%+v]", reflect.TypeOf(msg)))
+	}
+
 	event := cloudevents.NewEvent()
 	// CloudEvent specification: https://github.com/cloudevents/spec/blob/v1.0/spec.md#required-attributes
-	event.SetType(notificationType)
+	event.SetType(fmt.Sprintf("%v.%v", cloudEventTypePrefix, notificationType))
 	event.SetSource(cloudEventSource)
-	event.SetID(uuid.New().String())
-	event.SetTime(time.Now())
+	event.SetID(fmt.Sprintf("%v.%v", executionId, phase))
+	event.SetTime(eventTime)
 	event.SetExtension(jsonSchemaURLKey, jsonSchemaURL)
 
 	if err := event.SetData(cloudevents.ApplicationJSON, &msg); err != nil {
 		p.systemMetrics.PublishError.Inc()
-		logger.Errorf(ctx, "Failed to encode data with error: %v", err)
+		logger.Errorf(ctx, "Failed to encode message [%v] with error: %v", msg, err)
 		return err
 	}
 
 	if err := p.sender.Send(ctx, notificationType, event); err != nil {
 		p.systemMetrics.PublishError.Inc()
+		logger.Errorf(ctx, "Failed to send message [%v] with error: %v", msg, err)
 		return err
 	}
 	p.systemMetrics.PublishSuccess.Inc()
